Main program changed

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
-    public double calculate(double height, double weight) {
+    public int calculate(double height, double weight) {
         double result = weight / (height * height);
-        return result;
+        return (int) result;
     }
 }
